[Hyper]KnowlegeBaseRes/requirements.txt
sentence-transformers
faiss-cpu
numpy